Pass only the SourceSection instead of the node
Passing a node to another one is highly problematic. A node can only be part of the AST in a single position, because it has to maintain the proper parent relationship. And it can only have one parent.

So, to make things more clear, only pass it in when really necessary.

Signed-off-by: Stefan Marr <git@stefan-marr.de>