Added StorageAnalyzer to determine field offsets
This keeps the previous design for the object model, and gets the right objects, though, doesn’t need a complete rework as what we did in SOMns.

I think this solution here is better since it doesn’t introduce the extra indirection of a StorageAccessor.

This also added a UnsafeUtil to contain the code for getting to an unsafe instance.

Signed-off-by: Stefan Marr <git@stefan-marr.de>