Generalize the dispatch chain embedding to trivial methods
This removes the special casing for NewObjectPrim.

It is replaced by a mechanism to identify “trivial” methods, which can then be directly embedded in the dispatch chain.
This is currently supported for field accessors, global access, literals (except blocks with context), and NewObjectPrim.

To support setter methods, we need a new node as part of the sequence node, which makes sure the operation still returns self. So, this optimization checks this pattern and currently only supports setters that return self.

Signed-off-by: Stefan Marr <git@stefan-marr.de>