Introduce the JUMP2 bytecodes, which use the second byte to encode longer jumps
This is only needed for the Mandelbrot benchmark, I think.

The general idea is that we try to keep normal jumps as simple as possible.
By making all jumps using 3 bytes, and short jumps simply ignoring byte 3, we can keep our simple patch logic.

JUMP2* bytecodes now need to glue together the two bytes, but that should be relatively fast.

Signed-off-by: Stefan Marr <git@stefan-marr.de>