Do not store invokable in SArray anymore, but use linked hash map
This should be a general improvement for lookups, and memory use.
In the end, we had the invokable table anyway, now we just prepopulate it.

There’s one issue though, because the lookup does caching in the class hierarchy, we can’t just use the table to return the number of methods a class has.

This patch also changes how we know that a class has primitives, we now keep this as an explicit flag instead of iterating the invokables another time.

Signed-off-by: Stefan Marr <git@stefan-marr.de>