Remove second pass over bytecodes to determine stack hight
Instead keep track of the stack explicitly.
This spreads the knowledge a little more around, but it seems as if the table-based version I tried here caused a 5% slow down for the SomSom benchmarks, which is very strange.
So, let's see whether completely removing the second pass gives a speedup.

Signed-off-by: Stefan Marr <git@stefan-marr.de>