Add [Non]LocalFieldStringEqualsNode
This node inlines the argument and field reading and does the string equal.
There's theoretical an early abort for non-object fields.
Though, we don't have those in our benchmarks.
Well, perhaps uninitialized fields.
Though, the structure of the lookup, and caching makes that tricky. Not sure it's worth optimizing that.

Signed-off-by: Stefan Marr <git@stefan-marr.de>