Remove inlined blocks from literals and reuse literal index
This is a simple approach that simply relies on the blocks being the last literals added, which works for the supported inlinings, and because the parser just added them, before getting to the corresponding message send, where inlining is triggered.

So, we simply, from the back, remove the blocks just added.

Signed-off-by: Stefan Marr <git@stefan-marr.de>